Refactored java local notification code. Added cancelAll notifications.
